healthcare-rag/backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
python-multipart>=0.0.9
pydantic>=2.7.0
langchain>=0.1.13
langchain-community>=0.0.29
langchain-anthropic>=0.1.4
langchain-core>=0.1.33
chromadb>=0.4.24
sentence-transformers>=2.6.1
pypdf>=4.1.0
docx2txt>=0.8
tiktoken>=0.6.0
python-dotenv>=1.0.1
beautifulsoup4>=4.12.3
lxml>=5.1.0
